requirements.txt
requests==2.32.5
rich==14.2.0
python-dotenv==1.2.1
pyfiglet==1.0.4
packaging==25.0
